# docs/openapi.yaml
openapi: 3.0.3
info:
  title: LeagueLift API
  description: >
    LeagueLift revenue, fundraising, commerce, and fee-management API for youth sports
    organizations. This document covers Phase 0 (foundation), Phase 1 (organization
    onboarding, teams, tournaments, and public pages), Phase 2 (households,
    participants, fee templates, and fee assignments), Phase 3 slice 11
    (fundraising campaigns and their public pages) plus its remainder (campaign
    contribution recording via test-mode Stripe Checkout, real raised totals, and
    an inbound Stripe webhook receiver — DESIGN-DOC.md section 8.5; family
    credit rules remain a later slice pending section 19.3's open questions),
    traditional email/password authentication (superseding
    the originally planned external-IdP/Auth0 approach — see DESIGN-DOC.md section
    11.5), Phase 1's remaining file-upload/branding slice (organization logo/cover
    image via S3-compatible object storage — DESIGN-DOC.md section 11, ADR-012), and
    Phase 2's remainder (manual/offline fee payments, manual discounts/credits,
    org-wide collections listing and CSV export, and Stripe Connect Express
    onboarding-only scaffolding — DESIGN-DOC.md section 14.3, ADR-005 — no live
    charge routing yet), and Phase 4 slice 1 (apparel commerce proof: stores,
    Printify-catalog-backed products/variants with real Printify cost snapshots,
    a Printify catalog read proxy, multi-item test-mode Stripe Checkout with US
...
servers:
...
paths:
...
  /organizations:
    get:
      tags: [organizations]
      summary: List organizations the current user is a member of
      operationId: listOrganizations
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/SizeParam"
      responses:
        "200":
          description: Paginated list of organizations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizationPage"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [organizations]
      summary: Create an organization
      description: The creator is granted the OWNER membership role.
      operationId: createOrganization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrganizationRequest"
      responses:
        "201":
          description: Organization created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizationResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: Slug already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /organizations/{organizationId}/onboarding:
    parameters:
      - $ref: "#/components/parameters/OrganizationIdParam"
    get:
      tags: [organizations]
      summary: Get the organization's onboarding checklist status
      operationId: getOnboardingProgress
      responses:
        "200":
          description: Onboarding progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnboardingProgressResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # --- Payouts (Stripe Connect Express onboarding only — ADR-005, Phase 2 remainder) ---
...
  /me/dashboard-context:
    get:
      tags: [dashboard]
      summary: >
        Resolve which dashboard the signed-in user should land on (real role check —
        DESIGN-DOC.md section 10.1 — not a dev-only switcher). See
        DashboardContextService for the resolution order and its interim limitations.
      operationId: getDashboardContext
      responses:
        "200":
          description: Resolved dashboard context
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardContextResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # Every dashboard card is its own endpoint (DESIGN-DOC.md section 10.1/10.2), not one
  # bundled response per dashboard, so the frontend can load/error/empty each card
  # independently. Athlete endpoints are entirely demo data (no participant-login
  # concept exists yet); Owner/Coach/Parent endpoints are real where the schema
  # supports it and clearly tagged `isDemoData`/`isFundraisingDemoData`/`isRaisedDemoData`
  # where it does not — see the application services for the exact split.
  /me/dashboard/athlete/overview:
    get:
      tags: [dashboard]
      summary: Athlete dashboard — next event overview (demo)
      operationId: getAthleteOverview
      responses:
        "200":
          description: Athlete overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AthleteOverviewResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /me/dashboard/athlete/teams:
    get:
      tags: [dashboard]
      summary: Athlete dashboard — my teams (demo)
      operationId: getAthleteTeams
      responses:
        "200":
          description: Athlete's teams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AthleteTeamSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /me/dashboard/athlete/week-events:
    get:
      tags: [dashboard]
      summary: Athlete dashboard — this week's schedule (demo)
      operationId: getAthleteWeekEvents
      responses:
        "200":
          description: This week's events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScheduleItem"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /me/dashboard/athlete/recent-history:
    get:
      tags: [dashboard]
      summary: Athlete dashboard — recent history (demo)
      operationId: getAthleteRecentHistory
      responses:
        "200":
          description: Recent history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HistoryItem"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /me/dashboard/athlete/guardians:
    get:
      tags: [dashboard]
      summary: Athlete dashboard — guardians (demo)
      operationId: getAthleteGuardians
      responses:
        "200":
          description: Guardians
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GuardianSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /me/dashboard/athlete/orders:
    get:
      tags: [dashboard]
      summary: Athlete dashboard — orders (demo)
      operationId: getAthleteOrders
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /organizations/{organizationId}/dashboard/owner/summary:
    get:
      tags: [dashboard]
      summary: Owner dashboard — organization summary (real counts)
      operationId: getOwnerSummary
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
      responses:
        "200":
          description: Organization summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OwnerSummaryResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /organizations/{organizationId}/dashboard/owner/financial-overview:
    get:
      tags: [dashboard]
      summary: Owner dashboard — financial overview (demo)
      operationId: getOwnerFinancialOverview
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
      responses:
        "200":
          description: Financial overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancialOverviewResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /organizations/{organizationId}/dashboard/owner/attention-required:
    get:
      tags: [dashboard]
      summary: Owner dashboard — attention required (demo)
      operationId: getOwnerAttentionRequired
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
      responses:
        "200":
          description: Items needing attention
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttentionItem"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /organizations/{organizationId}/dashboard/owner/team-performance:
    get:
      tags: [dashboard]
      summary: Owner dashboard — team performance (real team/participant data; demo fundraising)
      operationId: getOwnerTeamPerformance
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
      responses:
        "200":
          description: Team performance rows
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeamPerformanceRow"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /organizations/{organizationId}/dashboard/owner/upcoming-events:
    get:
      tags: [dashboard]
      summary: Owner dashboard — upcoming events (demo)
      operationId: getOwnerUpcomingEvents
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
      responses:
        "200":
          description: Upcoming events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScheduleItem"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /organizations/{organizationId}/dashboard/owner/recent-activity:
    get:
      tags: [dashboard]
      summary: Owner dashboard — recent activity (real, from audit_event)
      operationId: getOwnerRecentActivity
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
      responses:
        "200":
          description: Recent activity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /organizations/{organizationId}/dashboard/owner/onboarding-progress:
    get:
      tags: [dashboard]
      summary: Owner dashboard — onboarding progress (demo)
      operationId: getOwnerOnboardingProgress
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
      responses:
        "200":
          description: Onboarding progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OwnerOnboardingProgress"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /organizations/{organizationId}/dashboard/owner/reports-snapshot:
    get:
      tags: [dashboard]
      summary: Owner dashboard — reports snapshot (demo)
      operationId: getOwnerReportsSnapshot
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
      responses:
        "200":
          description: Reports snapshot
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReportMetric"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /organizations/{organizationId}/dashboard/coach/teams:
    get:
      tags: [dashboard]
      summary: Coach dashboard — teams (real, scoped to the caller's actual assigned teams as of Phase 7/ADR-020 — not every team in the organization)
      operationId: getCoachTeams
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
      responses:
        "200":
          description: Teams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CoachTeamSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /organizations/{organizationId}/dashboard/coach/team-schedule:
    get:
      tags: [dashboard]
      summary: Coach dashboard — team schedule (demo)
      operationId: getCoachTeamSchedule
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
      responses:
        "200":
          description: Team schedule
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScheduleItem"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /organizations/{organizationId}/dashboard/coach/roster-summary:
    get:
      tags: [dashboard]
      summary: Coach dashboard — roster summary (real athlete count across the caller's assigned teams; demo attendance)
      operationId: getCoachRosterSummary
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
      responses:
        "200":
          description: Roster summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RosterSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /organizations/{organizationId}/dashboard/coach/team-page-status:
    get:
      tags: [dashboard]
      summary: Coach dashboard — team page status (real, when a page exists)
      operationId: getCoachTeamPageStatus
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
        - name: teamId
          in: query
          required: false
          description: Selects which of the caller's accessible teams to show (the team selector). Defaults to the alphabetically-first accessible team when omitted. 403s if the team isn't one the caller can access.
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Team page status, or null when the organization has no teams
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamPageStatusItem"
                nullable: true
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /organizations/{organizationId}/dashboard/coach/fundraising-progress:
    get:
      tags: [dashboard]
      summary: Coach dashboard — fundraising progress (real, including the raised amount, when the caller's team has an active campaign)
      operationId: getCoachFundraisingProgress
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
        - name: teamId
          in: query
          required: false
          description: Selects which of the caller's accessible teams to show (the team selector). Defaults to the alphabetically-first accessible team when omitted. 403s if the team isn't one the caller can access.
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Fundraising progress, or null when there is no active campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FundraisingProgress"
                nullable: true
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /organizations/{organizationId}/dashboard/coach/announcements:
    get:
      tags: [dashboard]
      summary: Coach dashboard — announcements (demo)
      operationId: getCoachAnnouncements
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
      responses:
        "200":
          description: Announcements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AnnouncementItem"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /organizations/{organizationId}/dashboard/coach/required-actions:
    get:
      tags: [dashboard]
      summary: Coach dashboard — required actions (demo)
      operationId: getCoachRequiredActions
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
      responses:
        "200":
          description: Required actions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RequiredActionItem"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /organizations/{organizationId}/households/{householdId}/dashboard/parent/overview:
    get:
      tags: [dashboard]
      summary: Parent dashboard — household overview (real)
      operationId: getParentOverview
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
        - $ref: "#/components/parameters/HouseholdIdParam"
      responses:
        "200":
          description: Household overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HouseholdOverviewResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /organizations/{organizationId}/households/{householdId}/dashboard/parent/athletes:
    get:
      tags: [dashboard]
      summary: Parent dashboard — athletes (real)
      operationId: getParentAthletes
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
        - $ref: "#/components/parameters/HouseholdIdParam"
      responses:
        "200":
          description: Athletes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AthleteSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /organizations/{organizationId}/households/{householdId}/dashboard/parent/family-schedule:
    get:
      tags: [dashboard]
      summary: Parent dashboard — family schedule (demo)
      operationId: getParentFamilySchedule
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
        - $ref: "#/components/parameters/HouseholdIdParam"
      responses:
        "200":
          description: Family schedule
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScheduleItem"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /organizations/{organizationId}/households/{householdId}/dashboard/parent/outstanding-balance:
    get:
      tags: [dashboard]
      summary: Parent dashboard — outstanding balance (real, approximate — see FeeLineItem)
      operationId: getParentOutstandingBalance
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
        - $ref: "#/components/parameters/HouseholdIdParam"
      responses:
        "200":
          description: Outstanding balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OutstandingBalance"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /organizations/{organizationId}/households/{householdId}/dashboard/parent/family-credits:
    get:
      tags: [dashboard]
      summary: Parent dashboard — family credits (demo)
      operationId: getParentFamilyCredits
      parameters:
        - $ref: "#/components/parameters/OrganizationIdParam"
        - $ref: "#/components/parameters/HouseholdIdParam"
      responses:
        "200":
          description: Family credits
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FamilyCredits"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /platform/dashboard/summary:
    get:
      tags: [dashboard]
      summary: Platform Admin dashboard — organization/user counts (real; new in Phase 7/ADR-020; requires platform.organization.view)
      operationId: getPlatformDashboardSummary
      responses:
        "200":
          description: Platform summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformSummaryResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    PageParam:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
...
security:
  - bearerAuth: []